muchos cambios

+++ src/usuarios/usuarios.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Param, Delete, UsePipes, ValidationPipe, Put, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Body, Param, Delete, UsePipes, ValidationPipe, Put, UseGuards, UnauthorizedException } from '@nestjs/common';
 import { UsuariosService } from './usuarios.service';
 import { CreateUsuarioDto } from './dto/create-usuario.dto';
 import { UpdateUsuarioDto } from './dto/update-usuario.dto';
@@ -13,13 +13,24 @@ export class UsuariosController {
 
   @Post()
   @UsePipes(new ValidationPipe())
-  create(@Body() createUserDto: CreateUsuarioDto) {
+  create(@Body() createUserDto: CreateUsuarioDto, @GetUserWithRole() userWithRole: UserWithRole) {
+
+    if (userWithRole.role == "user") {
+
+      throw new UnauthorizedException ({
+        statusCode: 401,
+        error: "No autorizado",
+        message: "Usuario no autorizado para realizar esta acción."
+      })
+
+    }
+
     return this.usuariosService.create(createUserDto);
   }
 
   @Get()
-  findAll(@GetUserWithRole() userWithRole: UserWithRole) {
-    console.log(userWithRole)
+  findAll() {
+
     return this.usuariosService.findAll();
   }
 
@@ -32,12 +43,38 @@ export class UsuariosController {
   update(
     @Param('id', ValidacionPipeIntPipe) id: number,
     @Body() updateUserDto: UpdateUsuarioDto,
+    @GetUserWithRole() userWithRole: UserWithRole
   ){
+
+
+
+
+    if (userWithRole.role == "user") {
+
+      throw new UnauthorizedException ({
+        statusCode: 401,
+        error: "No autorizado",
+        message: "Usuario no autorizado para realizar esta acción."
+      })
+
+    }
+
     return this.usuariosService.update(id, updateUserDto);
   }
 
   @Delete(':id')
-  remove(@Param('id', ValidacionPipeIntPipe) id: number) {
+  remove(@Param('id', ValidacionPipeIntPipe) id: number, @GetUserWithRole() userWithRole: UserWithRole) {
+
+    if (userWithRole.role == "user") {
+
+      throw new UnauthorizedException ({
+        statusCode: 401,
+        error: "No autorizado",
+        message: "Usuario no autorizado para realizar esta acción."
+      })
+
+    }
+
     return this.usuariosService.remove(id);
   }
 }
